Refactoring: Var is now a value class rather than a type alias for Int, leading to much more natural syntax in many parts of the code base

+++ src/main/scala/at/forsyte/harrsh/heapautomata/instances/ReachabilityAutomaton.scala
@@ -8,7 +8,7 @@ import at.forsyte.harrsh.seplog.Var
   * Created by jens on 3/29/17.
   */
 class ReachabilityAutomaton(override val numFV : Int, from : Var, to : Var, negate : Boolean = false) extends BaseReachabilityAutomaton(numFV) {
-  assert(from <= numFV && to <= numFV)
+  assert(from.toInt <= numFV && to.toInt <= numFV)
 
   override def isFinal(s : ReachabilityInfo) = s.rm.isReachable(from, to) != negate
   override val description = AutomatonTask.keywords.reach + "_" + numFV
